# openapi/api-gateway-chatbot.openapi.yaml
openapi: 3.0.3
info:
  title: SE121 API Gateway - Chatbot
  version: 1.0.0
  description: Public API Gateway endpoint for the AI Assistant chatbot.

servers:
  - url: http://localhost:4000
    description: Local API Gateway

tags:
  - name: Assistant
    description: AI Assistant chatbot endpoints exposed by the API Gateway.

paths:
  /api/v1/assistant/messages:
    post:
      tags:
        - Assistant
      summary: Send a message to the AI Assistant
      description: >
        Sends a user message to the chatbot through the API Gateway. The gateway
        resolves the authenticated user id from the Clerk bearer token, retrieves
        any server-side context needed for RAG, and forwards the internal request
        to chatbot-service.
      operationId: sendAssistantMessage
      security:
        - clerkBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantMessageRequest"
            examples:
              basic:
                summary: Basic assistant question
                value:
                  message: How does friend recommendation work?
      responses:
        "200":
          description: Assistant response generated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantRespondResponse"
              examples:
                success:
                  summary: Assistant response
                  value:
                    success: true
                    data:
                      reply: Friend recommendation suggests users that may be relevant based on allowed profile and relationship signals.
                      sources:
                        - type: help_doc
                          id: feature-friend-recommendation
                          title: Friend recommendation feature
                          source: assistant_docs
                          score: 0.89
                      suggestedActions: []
                      model: llama-3.3-70b-versatile
                      provider: groq
        "400":
          description: Invalid request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid Clerk bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Chatbot service returned an error or is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "504":
          description: Chatbot service request timed out.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/assistant/chat-history/{userId}:
    get:
      tags:
        - Assistant
      summary: Get assistant chat history by user
      description: >
        Returns persisted assistant chat history for a user using cursor pagination.
        Sorted by `created_at desc, id desc`.
      operationId: getAssistantHistoryByUser
      security:
        - clerkBearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: User id mapped to the single chatbot conversation.
          schema:
            type: string
        - name: page_size
          in: query
          required: false
          description: Number of messages per page.
          schema:
            type: integer
            minimum: 1
            example: 20
        - name: before_created_at
          in: query
          required: false
          description: Cursor timestamp (ISO-8601). Must be provided with `before_id`.
          schema:
            type: string
            format: date-time
            example: "2026-04-20T12:30:00.000Z"
        - name: before_id
          in: query
          required: false
          description: Cursor message id. Must be provided with `before_created_at`.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Chat history page loaded successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatbotHistoryResponse"
              examples:
                success:
                  value:
                    success: true
                    data:
                      items:
                        - id: 3ea8ba14-67db-4edc-9f2e-a66ad91dbb8f
                          conversation_id: 86159c45-6f8f-47c8-99ad-9ef6f4f2a644
                          user_id: user-123
                          role: user
                          content: friend recommendation là gì?
                          intent: help_doc
                          sources: []
                          metadata:
                            message_kind: user
                          created_at: "2026-04-20T12:29:40.012Z"
                      next_cursor_created_at: "2026-04-20T12:29:40.012Z"
                      next_cursor_id: 3ea8ba14-67db-4edc-9f2e-a66ad91dbb8f
                      has_more: true
        "400":
          description: Invalid cursor parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid Clerk bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Chatbot service returned an error or is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: History feature is disabled on chatbot service.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "504":
          description: Chatbot service request timed out.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - Assistant
      summary: Clear assistant chat history by user
      description: Deletes persisted chat messages and clears Redis session memory for the user.
      operationId: clearAssistantHistoryByUser
      security:
        - clerkBearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: User id mapped to the single chatbot conversation.
          schema:
            type: string
      responses:
        "200":
          description: Chat history cleared successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatbotClearHistoryResponse"
              examples:
                success:
                  value:
                    success: true
                    data:
                      deleted_count: 18
                      session_cleared: true
        "401":
          description: Missing or invalid Clerk bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Chatbot service returned an error or is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: History feature is disabled on chatbot service.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "504":
          description: Chatbot service request timed out.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    clerkBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk session JWT. Use the Authorization header as `Bearer <token>`.

  schemas:
    AssistantMessageRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          minLength: 1
          example: How does friend recommendation work?

    AssistantRespondResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/AssistantRespondData"

    AssistantRespondData:
      type: object
      required:
        - reply
        - sources
        - suggestedActions
        - model
        - provider
      properties:
        reply:
          type: string
          example: Friend recommendation suggests users that may be relevant based on allowed profile and relationship signals.
        sources:
          type: array
          items:
            $ref: "#/components/schemas/AssistantSource"
        suggestedActions:
          type: array
          items:
            $ref: "#/components/schemas/AssistantSuggestedAction"
        model:
          type: string
          example: llama-3.3-70b-versatile
        provider:
          type: string
          example: groq

    AssistantSource:
      type: object
      required:
        - type
        - id
      properties:
        type:
          type: string
          example: help_doc
        id:
          type: string
          example: feature-friend-recommendation
        title:
          type: string
          nullable: true
          example: Friend recommendation feature
        source:
          type: string
          nullable: true
          example: assistant_docs
        score:
          type: number
          format: float
          nullable: true
          example: 0.89

    AssistantSuggestedAction:
      type: object
      required:
        - type
        - label
        - payload
      properties:
        type:
          type: string
          example: open_search
        label:
          type: string
          example: Open search
        payload:
          type: object
          additionalProperties: true
          example:
            query: friend recommendation

    ChatbotHistoryResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/ChatbotHistoryData"

    ChatbotHistoryData:
      type: object
      required:
        - items
        - has_more
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ChatbotHistoryMessage"
        next_cursor_created_at:
          type: string
          format: date-time
          nullable: true
        next_cursor_id:
          type: string
          format: uuid
          nullable: true
        has_more:
          type: boolean
          example: false

    ChatbotHistoryMessage:
      type: object
      required:
        - id
        - conversation_id
        - user_id
        - role
        - content
        - sources
        - metadata
        - created_at
      properties:
        id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        user_id:
          type: string
          example: user-123
        role:
          type: string
          example: assistant
        content:
          type: string
        intent:
          type: string
          nullable: true
          example: help_doc
        sources:
          type: array
          items:
            $ref: "#/components/schemas/AssistantSource"
        metadata:
          type: object
          additionalProperties: true
          example:
            message_kind: assistant
        created_at:
          type: string
          format: date-time

    ChatbotClearHistoryResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          required:
            - deleted_count
            - session_cleared
          properties:
            deleted_count:
              type: integer
              minimum: 0
              example: 18
            session_cleared:
              type: boolean
              example: true

    ErrorResponse:
      oneOf:
        - type: string
          example: Chatbot service unavailable
        - type: object
          additionalProperties: true
          example:
            message: Unauthorized
            statusCode: 401